% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genD.R
\name{genD}
\alias{genD}
\title{Generate Bates and Watts D Matrix}
\usage{
genD(func, x, method.args = list(), ...)
}
\arguments{
\item{func}{A function whose first (vector) argument is a parameter vector;}

\item{x}{The parameter vector passed to \code{func};}

\item{method.args}{Arguments of the algorithm, used for the approximation algorithm;}

\item{...}{Additional arguments passed to \code{func}.}
}
\value{
A list of...
}
\description{
Generate a matrix of function derivative information.
}
\details{
The derivatives are calculated numerically using Richardson improvement...
}
\examples{
func <- function(x) c(x[1], x[1], x[2]^2)
xxx <- c(2, 2, 5)
RcppnumDeriv::genD(func, xxx)
}
