% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppnumDeriv-package.R
\docType{package}
\name{RcppnumDeriv}
\alias{RcppnumDeriv}
\title{Accurate and Fast Numerical Derivatives}
\description{
Description goes here...
}
\author{
Alessandro Gasparini (alessandro.gasparini@ki.se)
}
